moved some NIfTI functions out of utilities and in imaging class since NIfTI files are imaging specific
