+++ python/lib/utilities.py
@@ -9,8 +9,6 @@ import tarfile
 import scipy.io
 import numpy
 import lib.exitcode
-import numpy as np
-import nibabel as nib
 
 
 __license__ = "GPLv3"
@@ -142,40 +140,3 @@ def update_set_file_path_info(set_file, fdt_file):
     scipy.io.savemat(set_file, dataset, False)
 
 
-def get_nifti_image_length_parameters(nifti_filepath):
-    """
-    Get the NIfTI image length dimensions (x, y, z and time for 4D dataset).
-
-    :param nifti_filepath: path to the NIfTI file
-     :type nifti_filepath: str
-
-    :return: tuple with the length of each dimension of the NIfTI file
-     :rtype: tuple
-    """
-
-    img = nib.load(nifti_filepath)
-
-    # get the voxel/time length array of the image
-    length = img.shape
-
-    return length
-
-
-def get_nifti_image_step_parameters(nifti_filepath):
-    """
-    Get the NIfTI image step information (xstep, ystep, zstep and number of volumes
-    for 4D dataset)
-
-    :param nifti_filepath: path to the NIfTI file
-     :type nifti_filepath: str
-
-    :return: tuple with the step information for the NIfTI file
-     :rtype: tuple
-    """
-
-    img = nib.load(nifti_filepath)
-
-    # get the voxel step/time step of the image
-    step = img.header.get_zooms()
-
-    return step
